VS 2013 compatibility

+++ src/matrix/ft/SEM.hpp
@@ -36,7 +36,7 @@ E (const Matrix< std::complex<T> >& in, const Matrix< std::complex<T> >& sm,
    const Vector<size_t>& nx, const Vector<NFFT<T> >& ft) {
 	Matrix< std::complex<T> > out (nx[2],nx[1]);
 #pragma omp parallel for schedule (guided, 1)
-    for (size_t j = 0; j < nx[1]; ++j) {
+    for (int j = 0; j < nx[1]; ++j) {
         int k = omp_get_thread_num();
         Column (out, j, ft[k] * (resize(((nx[0] == 2) ? Slice (sm, j) : Volume (sm, j)),size(in)) * in));
     }
@@ -58,7 +58,7 @@ EH (const Matrix< std::complex<T> >& in, const Matrix< std::complex<T> >& sm,
     const Vector<size_t>& nx, const Vector<NFFT<T> >& ft) {
 	Matrix< std::complex<T> > out (size((nx[0] == 2) ? Slice (sm, 0) : Volume (sm, 0)));
 #pragma omp parallel for schedule (guided, 1)
-	for (size_t j = 0; j < nx[1]; ++j) {
+	for (int j = 0; j < nx[1]; ++j) {
         int k = omp_get_thread_num();
         out += ft[k] ->* Column (in,j) * conj((nx[0] == 2) ? Slice (sm, j) : Volume (sm, j));
     }
